fix: 커스텀 구분자 기능 수정 및 코드 분리
- 커스텀 구분자 기능이 제대로 수행되지 않던 문제 수정 중이다.
- 입력 처리와 계산 로직을 모듈화하여 separator 폴더로 분리
- Application, StringCalculator, Separator 클래스로 기능 구분
